requirements.txt
fastapi[all]
uvicorn
kafka-python
redis
requests
beautifulsoup4
transformers
torch
spacy
pydantic
python-socketio
lxml